requirements.txt
anthropic>=0.39.0
beautifulsoup4>=4.12.0
requests>=2.31.0
Pillow>=10.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
lxml>=4.9.0
